refactor liquidity pool into separate eth/sol files (no functionality changes)
